chore(Code quality): Remove commented lines that fudge the original roll result
The issue with fudging the roll result is that it is not obvious to the player that a feature of their weapon was activated, because that fact, especially for sharpness is hidden in the value of the dice.

Furthermore, it the dice roll values obviously don't follow their intended distribuition anymore which may lead to suspicions among players who overlook the sharpness attribute.

Signed-off-by: m-2squared <noname@example.com>