feat(Typestrict PoC): Rearrange and document file
Algorithm overview:
Assume an object with keys whose type is a foundry.data.field. Foundry uses this to inject these members into its data models. E.G. {isSleeping: new BooleanField({required:false,nullable:true}) becomes {isSleeping?:boolean|null} on the data model. We want to recreate that transition on type level. The generic type information, attained through careful construction of the DataFieldOptionType is imperative for this.

Therefore, we first go through all the keys in the schema and map the type of the foundry.data.field to the object that they are supposed to become on the data model. At the same time, use the REQ(=required) and NULL(=nullable) to check wether thee target type needs to be a union with `undefined` and `null`, respectively.
Once we've done that (recursively for Array and Schema types) we take the entire mapped type and make all members that have `undefined` as part of their types opional. This is to allow omitting the key when the type is written as a literal. (i.e. `const a :NoneRequiredSchema` = {}; should be allowed.

Signed-off-by: m-2squared <noname@example.com>