fix(NPCAttack): remove in-place merging.
Somehow, by actually altering the context we produced an edge case where the NPC Attack (and apparently only that one) would not be instantiated correctly on some browsers (firefox, foundry browser).
I removed the in-place attribute but suspect that now the statement is entirely pointless, as we don't process the return value. However I lack the time to test this and therefor leave this code here.

Signed-off-by: m-2squared <176928581+m-2squared@users.noreply.github.com>